Online Help: Move pydoc._start_server reimplementation
