+++ spyder/plugins/onlinehelp/pydoc_patch.py
@@ -1,0 +1,120 @@
+# -*- coding: utf-8 -*-
+#
+# Copyright © Spyder Project Contributors
+# Licensed under the terms of the MIT License
+# (see spyder/__init__.py for details)
+
+"""PyDoc patch"""
+
+
+def _start_server(urlhandler, hostname, port):
+    """
+    Start an HTTP server thread on a specific port.
+
+    This is a reimplementation of `pydoc._start_server` to handle connection
+    errors for 'do_GET'.
+
+    Taken from PyDoc: https://github.com/python/cpython/blob/3.7/Lib/pydoc.py
+    """
+    import http.server
+    import email.message
+    import select
+    import threading
+    import time
+
+    class DocHandler(http.server.BaseHTTPRequestHandler):
+
+        def do_GET(self):
+            """Process a request from an HTML browser.
+
+            The URL received is in self.path.
+            Get an HTML page from self.urlhandler and send it.
+            """
+            if self.path.endswith('.css'):
+                content_type = 'text/css'
+            else:
+                content_type = 'text/html'
+            self.send_response(200)
+            self.send_header(
+                'Content-Type', '%s; charset=UTF-8' % content_type)
+            self.end_headers()
+            try:
+                self.wfile.write(self.urlhandler(
+                    self.path, content_type).encode('utf-8'))
+            except ConnectionAbortedError:
+                # Needed to handle error when client closes the connection,
+                # for example when the client stops the load of the previously
+                # requested page. See spyder-ide/spyder#10755
+                pass
+
+        def log_message(self, *args):
+            # Don't log messages.
+            pass
+
+    class DocServer(http.server.HTTPServer):
+
+        def __init__(self, host, port, callback):
+            self.host = host
+            self.address = (self.host, port)
+            self.callback = callback
+            self.base.__init__(self, self.address, self.handler)
+            self.quit = False
+
+        def serve_until_quit(self):
+            while not self.quit:
+                rd, wr, ex = select.select([self.socket.fileno()], [], [], 1)
+                if rd:
+                    self.handle_request()
+            self.server_close()
+
+        def server_activate(self):
+            self.base.server_activate(self)
+            if self.callback:
+                self.callback(self)
+
+    class ServerThread(threading.Thread):
+
+        def __init__(self, urlhandler, host, port):
+            self.urlhandler = urlhandler
+            self.host = host
+            self.port = int(port)
+            threading.Thread.__init__(self)
+            self.serving = False
+            self.error = None
+
+        def run(self):
+            """Start the server."""
+            try:
+                DocServer.base = http.server.HTTPServer
+                DocServer.handler = DocHandler
+                DocHandler.MessageClass = email.message.Message
+                DocHandler.urlhandler = staticmethod(self.urlhandler)
+                docsvr = DocServer(self.host, self.port, self.ready)
+                self.docserver = docsvr
+                docsvr.serve_until_quit()
+            except Exception as e:
+                self.error = e
+
+        def ready(self, server):
+            self.serving = True
+            self.host = server.host
+            self.port = server.server_port
+            self.url = 'http://%s:%d/' % (self.host, self.port)
+
+        def stop(self):
+            """Stop the server and this thread nicely"""
+            self.docserver.quit = True
+            self.join()
+            # explicitly break a reference cycle: DocServer.callback
+            # has indirectly a reference to ServerThread.
+            self.docserver = None
+            self.serving = False
+            self.url = None
+
+    thread = ServerThread(urlhandler, hostname, port)
+    thread.start()
+    # Wait until thread.serving is True to make sure we are
+    # really up before returning.
+    while not thread.error and not thread.serving:
+        time.sleep(.01)
+    return thread
